Conflicts:
	simhalt.cc
	simversion.h

+++ simconvoi.cc
@@ -954,7 +954,7 @@ convoi_t::route_infos_t& convoi_t::get_route_infos()
 
 		// calc route infos
 		route_infos.set_count(route_count);
-		uint16 i = max(0, current_route_index - 2);
+		uint32 i = max(0, current_route_index - 2);
 
 		koord3d current_tile = route.position_bei(i);
 		convoi_t::route_info_t &start_info = route_infos.get_element(i);
